requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
python-dotenv==1.0.1
python-binance==1.0.19
streamlit-autorefresh==0.1.2
